Upgrade ESlint to v9
