--- .prettierrc.json
@@ -1,3 +1,0 @@
-{
-  "trailingComma": "none"
-}
