--- eslint.config.js
@@ -1,73 +1,0 @@
-const globals = require("globals");
-const {configs: eslintConfigs} = require("@eslint/js");
-const eslintPluginImport = require("eslint-plugin-import");
-const eslintPluginStylistic = require("@stylistic/eslint-plugin");
-
-const config = [
-  {
-    files: ["**/*.js"],
-    languageOptions: {
-      globals: {
-        ...globals.browser,
-        ...globals.node,
-        config: "readonly",
-        Log: "readonly",
-        Module: "readonly",
-        moment: "readonly",
-        WeatherObject: "readonly"
-      }
-    },
-    plugins: {
-      ...eslintPluginStylistic.configs["all-flat"].plugins,
-      import: eslintPluginImport
-    },
-    rules: {
-      ...eslintConfigs.all.rules,
-      ...eslintPluginImport.configs.recommended.rules,
-      ...eslintPluginStylistic.configs["all-flat"].rules,
-      "capitalized-comments": "off",
-      complexity: "off",
-      "consistent-this": "off",
-      "id-length": "off",
-      "init-declarations": "off",
-      "line-comment-position": "off",
-      "max-lines": ["warn", 1000],
-      "max-lines-per-function": ["warn", 500],
-      "max-params": ["warn", 4],
-      "max-statements": ["warn", 200],
-      "multiline-comment-style": "off",
-      "no-inline-comments": "off",
-      "no-magic-numbers": "off",
-      "no-negated-condition": "off",
-      "no-ternary": "off",
-      "one-var": "off",
-      "sort-keys": "off",
-      "@stylistic/array-element-newline": ["error", "consistent"],
-      "@stylistic/dot-location": ["error", "property"],
-      "@stylistic/function-call-argument-newline": ["error", "consistent"],
-      "@stylistic/indent": ["error", 2],
-      "@stylistic/lines-around-comment": "off",
-      "@stylistic/quote-props": ["error", "as-needed"],
-      "@stylistic/padded-blocks": ["error", "never"]
-    }
-  }
-];
-
-/*
- * Set debug to true for testing purposes.
- * Since some plugins have not yet been optimized for the flat config,
- * we will be able to optimize this file in the future. It can be helpful
- * to write the ESLint config to a file and compare it after changes.
- */
-const debug = false;
-
-if (debug === true) {
-  const FileSystem = require("node:fs");
-  FileSystem.writeFile("eslint-config-DEBUG.json", JSON.stringify(config, null, 2), (error) => {
-    if (error) {
-      throw error;
-    }
-  });
-}
-
-module.exports = config;
