+++ eslint.config.mjs
@@ -1,0 +1,84 @@
+import eslintPluginImportX from "eslint-plugin-import-x";
+import eslintPluginJs from "@eslint/js";
+import eslintPluginStylistic from "@stylistic/eslint-plugin";
+import globals from "globals";
+
+const config = [
+  eslintPluginJs.configs.all,
+  eslintPluginImportX.flatConfigs.recommended,
+  {
+    files: ["**/*.js"],
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+        ...globals.node,
+        Log: "readonly",
+        Module: "readonly",
+        WeatherObject: "readonly",
+        config: "readonly",
+        moment: "readonly"
+      }
+    },
+    plugins: {
+      ...eslintPluginStylistic.configs["all-flat"].plugins
+    },
+    rules: {
+      ...eslintPluginStylistic.configs["all-flat"].rules,
+      "@stylistic/array-element-newline": ["error", "consistent"],
+      "@stylistic/dot-location": ["error", "property"],
+      "@stylistic/function-call-argument-newline": ["error", "consistent"],
+      "@stylistic/indent": ["error", 2],
+      "@stylistic/lines-around-comment": "off",
+      "@stylistic/padded-blocks": ["error", "never"],
+      "@stylistic/quote-props": ["error", "as-needed"],
+      "capitalized-comments": "off",
+      complexity: "off",
+      "consistent-this": "off",
+      "id-length": "off",
+      "init-declarations": "off",
+      "line-comment-position": "off",
+      "max-lines": ["warn", 1000],
+      "max-lines-per-function": ["warn", 500],
+      "max-params": ["warn", 4],
+      "max-statements": ["warn", 200],
+      "multiline-comment-style": "off",
+      "no-inline-comments": "off",
+      "no-magic-numbers": "off",
+      "no-negated-condition": "off",
+      "no-ternary": "off",
+      "one-var": "off",
+      "sort-keys": "off"
+    }
+  },
+  {
+    files: ["**/*.mjs"],
+    languageOptions: {
+      ecmaVersion: "latest",
+      globals: {
+        ...globals.node
+      },
+      sourceType: "module"
+    },
+    plugins: {
+      ...eslintPluginStylistic.configs["all-flat"].plugins
+    },
+    rules: {
+      ...eslintPluginStylistic.configs["all-flat"].rules,
+      "@stylistic/array-element-newline": "off",
+      "@stylistic/dot-location": ["error", "property"],
+      "@stylistic/function-call-argument-newline": ["error", "consistent"],
+      "@stylistic/function-paren-newline": "off",
+      "@stylistic/indent": ["error", 2],
+      "@stylistic/padded-blocks": ["error", "never"],
+      "@stylistic/quote-props": ["error", "as-needed"],
+      "func-style": "off",
+      "max-lines-per-function": ["error", 100],
+      "no-magic-numbers": "off",
+      "no-undefined": "off",
+      "one-var": "off",
+      "prefer-destructuring": "off"
+    }
+  }
+];
+
+export default config;
