+++ package.json
@@ -22,13 +22,13 @@
     "moment": "^2.30.1"
   },
   "devDependencies": {
-    "@eslint/js": "^8.57.0",
-    "@stylistic/eslint-plugin": "^1.7.0",
-    "eslint": "^8.57.0",
-    "eslint-plugin-import": "^2.29.1",
-    "globals": "^14.0.0",
-    "markdownlint-cli": "^0.39.0",
-    "prettier": "^3.2.5"
+    "@eslint/js": "^9.12.0",
+    "@stylistic/eslint-plugin": "^2.9.0",
+    "eslint": "^9.12.0",
+    "eslint-plugin-import-x": "^4.3.1",
+    "globals": "^15.11.0",
+    "markdownlint-cli": "^0.42.0",
+    "prettier": "^3.3.3"
   },
   "scripts": {
     "lint": "eslint . && markdownlint . --ignore node_modules && prettier --check .",
