+++ prettier.config.mjs
@@ -1,0 +1,13 @@
+const config = {
+  overrides: [
+    {
+      files: "*.md",
+      options: {
+        parser: "markdown"
+      }
+    }
+  ],
+  trailingComma: "none"
+};
+
+export default config;
